Feature/support readonly (#85)
* Feature/headers interface (#83)

* Refactored models

* WIP

Working on Security headers admin screen

* Add headings UI

* update version details, notes

* dotnet bump

* merge to single package

* revert epi interface versions

* Added Readonly mode support

Fixed issues #79, #80, #81
Updated Documentation
