+++ src/Sample/DemoSite/appsettings.json
@@ -23,7 +23,11 @@
             "ShouldMatchAll": "false"
           }
         }
+      },
+      "DataAccess": {
+
       }
+
     }
   },
   "JhooseSecurity": {
